% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringRF.R
\name{clusteringRF}
\alias{clusteringRF}
\title{Unsupervised Random Forest for clustering}
\usage{
clusteringRF(
  X,
  ntrees = 500,
  mtry = ncol(X),
  distance = "co-clustering",
  pruning = TRUE,
  weighting = FALSE,
  ncores = parallel::detectCores() - 1
)
}
\arguments{
\item{X}{a data.frame to be clustered}

\item{ntrees}{number of \code{divlust} trees. Default is \code{500}.}

\item{mtry}{number of variables selected at each \code{divlust} tree node.
Default is \code{ncol(X)}.}

\item{distance}{a character string, either "co-clustering" or "inertia".}

\item{pruning}{Logical (TRUE or FALSE). If TRUE, inertia trees are pruned.}

\item{weighting}{Logical (TRUE or FALSE). If TRUE, inertia nodes are weighted
in the calculation of variable importance.}

\item{ncores}{number of cpus to parallelize over. Default is
\code{parallel::detectCores()-1}.}
}
\value{
a list of length \code{n_trees}, each elements containing 3 matrices.
}
\description{
main wrapper
}
\examples{
library(palmerpenguins)
mypeng <- as.data.frame(penguins)
mypeng$year <- factor(as.character(mypeng$year),
                         levels=c("2007", "2008", "2009"),
                         ordered=TRUE)
set.seed(123)
forest_clust <- clusteringRF(na.omit(mypeng[mypeng$sex=="male", -c(1, 7)]), ntrees = 50, ncores = 1)

if(interactive()){
 resume <- summary(forest_clust)

 plot(resume$Matrix)
}

if(interactive()){
   ncol <- 15
   nobs_pg <- 43
   diff_mean <- 5
   X <- matrix(rnorm(n = ncol*nobs_pg, mean = 0, sd = 1), ncol=ncol)
   Y <- matrix(rnorm(n = ncol*nobs_pg, mean = diff_mean, sd = 1), ncol=ncol)
   mysim_data <- rbind.data.frame(X,Y)
   res <- clusteringRF(mysim_data)
   s <- summary(res)
   plot(s)
}
}
