% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{make_CSI_metrics_coclust}
\alias{make_CSI_metrics_coclust}
\title{Computes Clustering Stability Importance (CSI) Metrics for co-clustering distance}
\usage{
make_CSI_metrics_coclust(X_oob, path_cluster, prediction)
}
\arguments{
\item{X_oob}{a data frame of out-of-bag observations}

\item{path_cluster}{a data frame produced by \code{make_path_cluster()}}

\item{prediction}{a data frame produced by \code{make_prediction()}}
}
\value{
a list with 2 elements : Out-of-bag Reassignment Rate (CSI_ORR) which counts
the proportion of observations whose cluster assignment changed after
permutation and ARI Stability Importance (CSI_ARI) which computes adjusted rand index between the
original and permuted cluster assignments.
}
\description{
Function which estimates the importance of each variable on the clustering
using a permuation approach.
}
