% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{make_MDA_importance}
\alias{make_MDA_importance}
\title{Computes Mean Decrease of Accuracy (MDA) importance metrics}
\usage{
make_MDA_importance(X_oob, path_cluster, prediction)
}
\arguments{
\item{X_oob}{a data frame of out-of-bag observations}

\item{path_cluster}{a data frame produced by \code{make_path_cluster()}}

\item{prediction}{a data frame produced by \code{make_prediction()}}
}
\value{
a list with 2 elements : permutation_count which counts
the proportion of observations whose cluster assignment changed after
permutation and ARI_permutation which computes adjusted rand index between the
original and permuted cluster assignments.
}
\description{
Function which estimates the importance of each variable using a
permuation approach.
}
