% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{make_prediction}
\alias{make_prediction}
\title{Predicts cluster membership for our-of-bag observations}
\usage{
make_prediction(X_oob, path_cluster)
}
\arguments{
\item{X_oob}{a data frame of out-of-bag observations}

\item{path_cluster}{a data frame produced by \code{make_path_cluster()}}
}
\value{
a data frame with the predicted cluster label for each observation
}
\description{
Function which assigns each observation to the right cluster.
}
