% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.clusteringRF.R
\name{summary.clusteringRF}
\alias{summary.clusteringRF}
\title{summary S3 method for clusteringRF objects}
\usage{
\method{summary}{clusteringRF}(object, ...)
}
\arguments{
\item{object}{the results of the apply function on the RF function (tree)}

\item{...}{more parameters if required}
}
\value{
Provides the cumulative similarity matrix and some analysis
}
\description{
Merge all matrices and some analysis
}
