% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{tree}
\alias{tree}
\title{Computes a divclust tree}
\usage{
tree(
  X,
  mtry = ncol(X),
  distance = c("co-clustering"),
  pruning = TRUE,
  weighting = FALSE
)
}
\arguments{
\item{X}{data.frame of input data}

\item{mtry}{number of variables selected at each tree node}

\item{distance}{a character string, either "co-clustering" or "inertia".}

\item{pruning}{Logical (TRUE or FALSE). If TRUE, the inertia tree is pruned.}

\item{weighting}{Logical (TRUE or FALSE). If TRUE, inertia nodes are weighted
in the calculation of variable importance.}
}
\value{
dissimilarity matrix and information on the importance of variables
}
\description{
Computes a divisive monothetic tree from the divclust package using either
co-clustering or inertia distances
}
