% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{make_CS_metrics_inertia}
\alias{make_CS_metrics_inertia}
\title{Computes Clustering Stability (CS) Metrics for inertia distance}
\usage{
make_CS_metrics_inertia(X_oob, path_cluster, prediction, dist_clusters)
}
\arguments{
\item{X_oob}{a data frame of out-of-bag observations}

\item{path_cluster}{a data frame produced by \code{make_path_cluster()}}

\item{prediction}{a data frame produced by \code{make_prediction()}}

\item{dist_clusters}{a distance matrix}
}
\value{
a list with 1 element : Inertia Permutation (CS_IP) which computes the sum
of inertia between the original and permuted clusters assignments.
}
\description{
Function which estimates the importance of each variable on the clustering
using a permutation approach.
}
