% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{make_path_cluster}
\alias{make_path_cluster}
\title{Constructs the path rules for each cluster}
\usage{
make_path_cluster(description)
}
\arguments{
\item{description}{monothetic description of the leaves}
}
\value{
a data frame with six columns : the cluster label, the variable name, the lower
and upper bounds of the variable's range (for quantitative variables),
the categories (for qualitative variables), and the variable type
}
\description{
Function which builds a data frame describing, for every cluster,
the associated variable intervals.
}
